+++ project/src/main/resources/static/posizioni.js
@@ -41,15 +41,17 @@ export default Vue.component("posizioni", {
     },
     async deletePosizione(id) {
       this.$emit("caricamento", true);
-      const res = await (
-        await fetch(
-          "/api/" + this.utente.tipo + "/eliminaPosizione?idPosizione=" + id,
-          {
-            method: "DELETE",
-          }
-        )
-      ).text();
-      this.$emit("notifica", res);
+      if(confirm("Sei sicuro di voler eliminare la posizione?")) {
+        const res = await (
+          await fetch(
+            "/api/" + this.utente.tipo + "/eliminaPosizione?idPosizione=" + id,
+            {
+              method: "DELETE",
+            }
+          )
+        ).text();
+        this.$emit("notifica", res);
+      }
       this.$emit("caricamento", false);
       await this.aggiorna();
     },
